+++ packages/next-common/components/overview/overview.jsx
@@ -4,7 +4,7 @@ import { useChainSettings } from "next-common/context/chain";
 import TreasuryStats from "./treasuryStats";
 import WithPallet from "next-common/components/common/withPallet";
 import FellowshipSalaryOverview from "next-common/components/overview/fellowship/salary/overview";
-import MembersInduction from "./membersInduction";
+import MembersInduction from "./fellowship/membersInduction";
 import { useChain } from "next-common/context/chain";
 import Chains from "next-common/utils/consts/chains";
 export default function Overview() {
